pyproject.toml
[project]
name = "compass"
version = "0.1.6"
description = "COMputational PAthology and Spatial Statistics"
readme = "README.md"
authors = [
    { name = "vladpopovici", email = "popovici@bioxlab.org" }
]
requires-python = ">=3.12"
dependencies = [
    "bottleneck>=1.6.0",
    "dask-image>=2025.11.0",
    "dask[array,complete,dataframe]>=2024.0.0",
    "fastapi[standard]>=0.115.12",
    "geojson>=3.1.0",
    "h5netcdf>=1.7.3",
    "h5py>=3.13.0",
    "hdf5plugin>=6.0.0",
    "joblib>=1.4.2",
    "mahotas>=1.4.18",
    "mkdocs>=1.6.1",
    "mkl>=2025.0.1",
    "netcdf4>=1.7.3",
    "numbagg>=0.9.3",
    "opencv-python>=4.10.0.84",
    "openslide-bin>=4.0.0.6",
    "openslide-python>=1.4.1",
    "orjson>=3.10.16",
    "pydantic>=2.11.4",
    "pydantic-settings>=2.9.1",
    "pyvips>=2.2.3",
    "scikit-image>=0.24.0",
    "scikit-learn>=1.6.0",
    "scipy>=1.14.0",
    "shapely>=2.0.6",
    "simplejson>=3.19.3",
    "tqdm>=4.67.1",
    "wsitk-annot",
    "xarray>=2025.12.0",
    "xmltodict>=0.14.2",
    "zarr>=2.18.3",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "matplotlib>=3.10.1",
    "mkdocs-material>=9.6.14",
    "mkdocstrings[python]>=0.29.1",
]

[tool.uv.sources]
wsitk-annot = { git = "https://github.com/vladpopovici/wsitk-annot" }
